Inline all modules into main
The tool is very small, might as well define everything we need in
main.rs
