+++ src/main.rs
@@ -1,16 +1,30 @@
-use anyhow::Result;
+use anyhow::{Context, Result};
+use futures::{AsyncRead, AsyncWrite};
+use libp2p::core::identity::ed25519::Keypair;
+use libp2p::core::muxing::StreamMuxerBox;
+use libp2p::core::transport::Boxed;
+use libp2p::core::upgrade::{SelectUpgrade, Version};
 use libp2p::dns::TokioDnsConfig;
 use libp2p::futures::StreamExt;
+use libp2p::identity::ed25519::SecretKey;
+use libp2p::mplex::MplexConfig;
+use libp2p::noise::{NoiseConfig, X25519Spec};
+use libp2p::ping::{Ping, PingConfig, PingEvent};
 use libp2p::rendezvous::{Config, Event as RendezvousEvent, Rendezvous};
 use libp2p::swarm::{SwarmBuilder, SwarmEvent};
 use libp2p::tcp::TokioTcpConfig;
-use libp2p::{identity, PeerId, Transport};
-use rendezvous_server::tracing::init;
-use rendezvous_server::transport::authenticate_and_multiplex;
-use rendezvous_server::{generate_secret_key_file, load_secret_key_from_file, Behaviour, Event};
-use std::path::PathBuf;
+use libp2p::yamux::YamuxConfig;
+use libp2p::{identity, noise, rendezvous, Multiaddr, PeerId, Transport};
+use std::fmt;
+use std::path::{Path, PathBuf};
+use std::time::Duration;
 use structopt::StructOpt;
+use tokio::fs;
+use tokio::fs::{DirBuilder, OpenOptions};
+use tokio::io::AsyncWriteExt;
 use tracing::level_filters::LevelFilter;
+use tracing_subscriber::fmt::time::ChronoLocal;
+use tracing_subscriber::FmtSubscriber;
 
 #[derive(Debug, StructOpt)]
 struct Cli {
@@ -37,7 +51,7 @@ struct Cli {
 async fn main() -> Result<()> {
     let cli = Cli::from_args();
 
-    init(LevelFilter::INFO, cli.json, cli.timestamp);
+    init_tracing(LevelFilter::INFO, cli.json, cli.timestamp);
 
     let secret_key = match cli.generate_secret {
         true => generate_secret_key_file(cli.secret_file).await?,
@@ -87,7 +101,7 @@ async fn main() -> Result<()> {
                 SwarmEvent::Behaviour(Event::Rendezvous(RendezvousEvent::RegistrationExpired(
                     registration,
                 ))) => {
-                    tracing::info!(peer=%registration.record.peer_id(), namespace=%registration.namespace, addresses=%rendezvous_server::tracing::Addresses(registration.record.addresses()), ttl=registration.ttl, "Registration expired");
+                    tracing::info!(peer=%registration.record.peer_id(), namespace=%registration.namespace, addresses=%Addresses(registration.record.addresses()), ttl=registration.ttl, "Registration expired");
                 }
                 SwarmEvent::Behaviour(Event::Rendezvous(RendezvousEvent::PeerUnregistered {
                     peer,
@@ -105,4 +119,147 @@ async fn main() -> Result<()> {
             }
         }
     }
+}
+
+#[derive(Debug)]
+pub enum Event {
+    Rendezvous(rendezvous::Event),
+    Ping(PingEvent),
+}
+
+impl From<rendezvous::Event> for Event {
+    fn from(event: rendezvous::Event) -> Self {
+        Event::Rendezvous(event)
+    }
+}
+
+impl From<PingEvent> for Event {
+    fn from(event: PingEvent) -> Self {
+        Event::Ping(event)
+    }
+}
+
+#[derive(libp2p::NetworkBehaviour)]
+#[behaviour(event_process = false)]
+#[behaviour(out_event = "Event")]
+pub struct Behaviour {
+    ping: Ping,
+    rendezvous: Rendezvous,
+}
+
+impl Behaviour {
+    pub fn new(rendezvous: Rendezvous) -> Self {
+        Self {
+            // TODO: Remove Ping behaviour once https://github.com/libp2p/rust-libp2p/issues/2109 is fixed
+            // interval for sending Ping set to 24 hours
+            ping: Ping::new(
+                PingConfig::new()
+                    .with_keep_alive(false)
+                    .with_interval(Duration::from_secs(86_400)),
+            ),
+            rendezvous,
+        }
+    }
+}
+
+pub async fn load_secret_key_from_file(path: impl AsRef<Path> + fmt::Debug) -> Result<SecretKey> {
+    let bytes = fs::read(&path)
+        .await
+        .with_context(|| format!("No secret file at {:?}", path))?;
+    let secret_key = SecretKey::from_bytes(bytes)?;
+    Ok(secret_key)
+}
+
+pub async fn generate_secret_key_file(path: PathBuf) -> Result<SecretKey> {
+    if let Some(parent) = path.parent() {
+        DirBuilder::new()
+            .recursive(true)
+            .create(parent)
+            .await
+            .with_context(|| format!("Could not create directory for secret file: {:?}", parent))?;
+    }
+    let mut file = OpenOptions::new()
+        .write(true)
+        .create_new(true)
+        .open(&path)
+        .await
+        .with_context(|| format!("Could not generate secret file at {:?}", &path))?;
+
+    let keypair = Keypair::generate();
+    let secret_key = SecretKey::from(keypair);
+
+    file.write_all(secret_key.as_ref()).await?;
+
+    Ok(secret_key)
+}
+
+pub fn init_tracing(level: LevelFilter, json_format: bool, timestamp: bool) {
+    if level == LevelFilter::OFF {
+        return;
+    }
+
+    let is_terminal = atty::is(atty::Stream::Stderr);
+
+    let builder = FmtSubscriber::builder()
+        .with_env_filter(format!("rendezvous_server={}", level))
+        .with_writer(std::io::stderr)
+        .with_ansi(is_terminal)
+        .with_timer(ChronoLocal::with_format("%F %T".to_owned()))
+        .with_target(false);
+
+    if json_format {
+        builder.json().init();
+        return;
+    }
+
+    if !timestamp {
+        builder.without_time().init();
+        return;
+    }
+    builder.init();
+}
+
+pub struct Addresses<'a>(pub &'a [Multiaddr]);
+
+// Prints an array of multiaddresses as a comma seperated string
+impl fmt::Display for Addresses<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let display = self
+            .0
+            .iter()
+            .map(|addr| addr.to_string())
+            .collect::<Vec<String>>()
+            .join(",");
+        write!(f, "{}", display)
+    }
+}
+
+/// "Completes" a transport by applying the authentication and multiplexing
+/// upgrades.
+///
+/// Even though the actual transport technology in use might be different, for
+/// two libp2p applications to be compatible, the authentication and
+/// multiplexing upgrades need to be compatible.
+pub fn authenticate_and_multiplex<T>(
+    transport: Boxed<T>,
+    identity: &identity::Keypair,
+) -> Result<Boxed<(PeerId, StreamMuxerBox)>>
+where
+    T: AsyncRead + AsyncWrite + Unpin + Send + 'static,
+{
+    let auth_upgrade = {
+        let noise_identity = noise::Keypair::<X25519Spec>::new().into_authentic(identity)?;
+        NoiseConfig::xx(noise_identity).into_authenticated()
+    };
+    let multiplex_upgrade = SelectUpgrade::new(YamuxConfig::default(), MplexConfig::new());
+
+    let transport = transport
+        .upgrade(Version::V1)
+        .authenticate(auth_upgrade)
+        .multiplex(multiplex_upgrade)
+        .timeout(Duration::from_secs(20))
+        .map(|(peer, muxer), _| (peer, StreamMuxerBox::new(muxer)))
+        .boxed();
+
+    Ok(transport)
 }
